// ClientApp/src/app/physician/models/user.ts
export class User {
  Id: number;
  Name: string;
  Email: string;
  CreateDate: Date;
  Role: string;
  PhoneNumber: string;
  PreferredCommunicationMethod: string;
  PractitionerID: number;
  IsImpersonating: boolean;
  ConfirmedDate: Date;
  ConfirmedByUserID: number;
  RequiresPayment: boolean;
  CustomerId: string;
  PaymentId: string;
  ActiveAppointmentCount: number;
}